Disable binlogging by default (#204)
* Add binlogging caps

The default binlogging for mariadb will easily fill a modest disk after a couple of weeks on a busy instance.  This setting will limit that usage and have mariadb rotate the logs out after 1 day.

* Disable binlogging on mariadb

Binlogging will, by default, run amok on the drive and fill it.  It's also going to run a lot of write cycles to SD cards which will considerably shorten their lifespan.

* Update run.sh

* Update config.json